Sprint 4

+++ src/js/eventHandler.mjs
@@ -5,17 +5,21 @@ const eventHandler = (button) => {
     const li = button.closest('.todo-item');
     // получим значение id элемента в переменную
     const id = Number(li.getAttribute('id'));
+    // защита от ошибок
+    if (!id) return;
 
     if (button.classList.contains('delete-btn')) {
         storage.deleteTask(id);
-        li.remove(); // удалить задачу
+        // удалить задачу
+        li.remove();
     }
 
     if (button.classList.contains('complete-btn')) {
         const li = button.closest('li');
         const id = parseInt(li.id, 10);
         storage.toggleTask(id);
-        li.classList.toggle('completed'); // визуальное изменение
+        // визуальное изменение
+        li.classList.toggle('completed');
     }
 };
